requirements.txt
###### Requirements without Version Specifiers ######
atlassian-python-api